// packages/core-sdk/src/enums/ChainDisplayName.ts
export enum ChainDisplayName {
	ETHEREUM = 'Ethereum',
	MOONRIVER = 'Moonriver',
	MOONBEAM = 'Moonbeam',
	RINKEBY = 'Rinkeby',
	MOONBASE = 'MoonBase',
	BOBA = 'Boba',
	BOBABEAM = 'BobaBeam',
	BOBA_RINKEBY = 'Boba Rinkeby',
	BOBABASE = 'BobaBase',
	AURORA = 'Aurora',
	POLYGON = 'Polygon',
	FANTOM = 'Fantom',
	BOBAOPERA = 'BobaOpera',
	AVALANCHE = 'Avalanche'
}